refactor(core/stellar): move generic Soroban layout helpers to layout.py.
SCVal formatting and invocation confirmation helpers are used by both
the transaction flow and the standalone authorization signing flow, so
they are no longer operation-specific.

[no changelog]
